103 complete backend documentation

+++ backend/src/router/actionReaction/actionReaction.js
@@ -8,9 +8,9 @@ exports.postArea = async (req, res) => {
         if (!action || !reaction)
             return api_formatter(req, res, 400, "error", "Missing required fields", null);
         if (action.service == "github" && req.user.github_token == null)
-            return api_formatter(req, res, 400, "error", "You need to connect your github account to use github's action", null);
+            return api_formatter(req, res, 401, "error", "You need to connect your github account to use github's action", null);
         if (reaction.service == "discord" && req.user.discord_token == null)
-            return api_formatter(req, res, 400, "error", "You need to connect your discord account to use discord's reaction", null);
+            return api_formatter(req, res, 401, "error", "You need to connect your discord account to use discord's reaction", null);
 
         const tokens = `{
             "${action.service}": "${action.service == "github" ? req.user.github_token.access_token : ""}",
@@ -30,14 +30,15 @@ exports.postArea = async (req, res) => {
     }
     catch (err) {
         console.error(err);
-        return api_formatter(req, res, 500, "error", "An error occured while trying to get the discord token", null, err);
+        return api_formatter(req, res, 500, "error", "An error occured while trying to create the area", null, err);
     }
 };
 
 exports.getArea = async (req, res) => {
     try {
         const actionReactions = await ActionReactionModel.find({ creator_id: req.user.unique_id });
-        console.log(actionReactions);
+        if (actionReactions.length === 0)
+            return api_formatter(req, res, 404, "notFound", "Action Reaction not found");
         const parsedData = actionReactions.map(actionReaction => ({
             id: actionReaction.unique_id,
             action: actionReaction.Action,
@@ -47,7 +48,7 @@ exports.getArea = async (req, res) => {
     }
     catch (err) {
         console.error(err);
-        return api_formatter(req, res, 500, "error", "An error occured while trying to get the discord token", null, err);
+        return api_formatter(req, res, 500, "error", "An error occured while trying to get the user's area", null, err);
     }
 };
 
@@ -56,6 +57,8 @@ exports.deleteArea = async (req, res) => {
         const { id } = req.body;
         if (!id)
             return api_formatter(req, res, 400, "error", "Missing required fields", null);
+        if (!await ActionReactionModel.exists({ unique_id: id }))
+            return api_formatter(req, res, 404, "notFound", "Action Reaction not found");
         await ActionReactionModel.deleteOne({ unique_id: id });
         return api_formatter(req, res, 200, "success", "Action Reaction deleted");
     }
